@@ -1,6 +1,6 @@
 {
   "name": "serverless-domain-manager",
-  "version": "3.3.2",
+  "version": "4.2.2",
   "engines": {
     "node": ">=4.0"
   },
@@ -27,16 +27,16 @@
   "main": "dist/index.js",
   "bin": {},
   "scripts": {
-    "test": "tsc --project . && nyc mocha -r ts-node/register test/unit-tests/index.test.ts && nyc report --reporter=text-summary",
-    "integration-test": "node ./node_modules/istanbul/lib/cli.js test _mocha test/integration-tests -- -R spec",
-    "lint": "tslint --project .",
+    "test": "nyc mocha -r ts-node/register --project tsconfig.tests.json test/unit-tests/index.test.ts && nyc report --reporter=text-summary",
+    "integration-test": "nyc mocha -r ts-node/register --project tsconfig.tests.json test/integration-tests/*.test.ts && nyc report --reporter=text-summary",
+    "lint": "tslint --project . && tslint --project tsconfig.tests.json",
     "build": "tsc --project ."
   },
   "files": [
     "*.js",
     "*.ts",
     "*.json",
-    "dist/*.js"
+    "dist"
   ],
   "nyc": {
     "extension": [
@@ -44,29 +44,30 @@
     ]
   },
   "devDependencies": {
-    "@types/chai": "^4.1.7",
-    "@types/chai-spies": "^1.0.0",
-    "@types/mocha": "^5.2.5",
-    "@types/node": "^12.0.0",
+    "@types/chai": "^4.2.12",
+    "@types/chai-spies": "^1.0.2",
+    "@types/mocha": "^5.2.7",
+    "@types/node": "^12.12.58",
     "aws-sdk-mock": "^1.7.0",
     "chai": "^3.5.0",
     "chai-spies": "^1.0.0",
     "istanbul": "^0.4.5",
-    "mocha": "^5.2.0",
-    "mocha-param": "^2.0.0",
+    "mocha": "^7.2.0",
+    "mocha-param": "^2.0.1",
     "nyc": "^14.1.1",
     "randomstring": "^1.1.5",
-    "request": "^2.88.0",
-    "request-promise-native": "^1.0.5",
-    "shelljs": "^0.8.3",
-    "ts-node": "^8.0.1",
-    "tslint": "^5.1.0",
-    "tslint-microsoft-contrib": "^6.0.0",
-    "typescript": "^3.2.2",
+    "request": "^2.88.2",
+    "request-promise-native": "^1.0.9",
+    "serverless": "^1.83.0",
+    "shelljs": "^0.8.4",
+    "ts-node": "^8.10.2",
+    "tslint": "^5.20.1",
+    "tslint-microsoft-contrib": "^6.2.0",
+    "typescript": "^3.9.7",
     "wrappy": "^1.0.2"
   },
   "dependencies": {
-    "aws-sdk": "^2.490.0",
-    "chalk": "^2.4.1"
+    "aws-sdk": "^2.752.0",
+    "chalk": "^2.4.2"
   }
 }
